CHROMIUM: automatically mount /run at boot
We have a small pre-startup shell script that initializes /run and
then processes selinux labels.  This 2 syscalls (and libselinux).
Since we consider /run to be absolutely required as part of the
system, and we always want it to be tmpfs, move the setup to
upstart itself to simplify.

BUG=chromium:1063545
TEST=CQ passes

Change-Id: Ic3f8545a8fdaee37eb6de59f95c5a16b8a758304
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/upstart/+/2546468
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>